# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fiber API with OAuth2 password flow",
        "title": "Openapi spec",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/user-session/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a list of users with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Retrieves all users with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserSessionExtendedPaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user-session/{id}/detail/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieve single user session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve single user session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserSessionExtendedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user-session/{id}/revoke/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a list of users with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Retrieves all user sessions with pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessagedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ChoiceBase": {
            "type": "object",
            "required": [
                "label",
                "value"
            ],
            "properties": {
                "label": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "MessagedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "PermissionDeniedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Permission denied"
                }
            }
        },
...
        "UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Unauthorized - Invalid token"
                }
            }
        },
...
        "UserResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "isActive",
                "role",
                "username"
            ],
            "properties": {
                "carPark": {
                    "description": "optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ChoiceBase"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "fullName": {
                    "description": "required in TS, can be empty",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/ChoiceBase"
                },
                "updatedAt": {
                    "description": "optional, can be nil",
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "UserSessionExtendedPaginatedResponse": {
            "type": "object",
            "required": [
                "limit",
                "page",
                "rows"
            ],
            "properties": {
                "limit": {
                    "description": "items per page",
                    "type": "integer"
                },
                "page": {
                    "description": "current page",
                    "type": "integer"
                },
                "rows": {
                    "description": "generic slice of any type",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/UserSessionExtendedResponse"
                    }
                }
            }
        },
        "UserSessionExtendedResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "expireAt",
                "id"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expireAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "ipAddress": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "macPassword": {
                    "type": "string",
                    "example": "mac-password"
                },
                "macUsername": {
                    "type": "string",
                    "example": "mac-username"
                },
                "revokedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/UserResponse"
                },
                "userAgent": {
                    "type": "string",
                    "example": "Mozilla/5.0"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/v1/auth/login",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}